+++ services/offline/src/index.ts
@@ -1,6 +1,8 @@
 export { OfflineProvider } from './lib/offline-provider'
 export { CacheableSection, useCacheableSection } from './lib/cacheable-section'
 export { useCachedSections } from './lib/cacheable-section-state'
-export { useOnlineStatus } from './lib/online-status'
+// Use "useOnlineStatus" name for backwards compatibility
+export { useNetworkStatus as useOnlineStatus } from './lib/network-status'
 export { useOnlineStatusMessage } from './lib/online-status-message'
 export { clearSensitiveCaches } from './lib/clear-sensitive-caches'
+export { useDhis2ConnectionStatus } from './lib/dhis2-connection-status'